Fix #76 - Normalize satellite color shading across SNR and C/N0 for Android 7.0 and higher
* Sky view on Android 7.0 and higher - To determine color shading (red->green), SNR and C/N0 will both use the range of 10 (fully red) to 45 (fully green). If someone toggles between viewing SNR and C/N0 (using the "Use GNSS APIs" setting), this will result in more intuitive experience where the color shading for a given numeric value remains consistent between SNR and C/N0.
* Sky view on Android versions below 7.0 - Nothing will change - color shading for SNR (C/N0 isn't available) will continue to use the range of 0-30 for SNR.

The reasoning for 3) is that I don't want to suddenly change the behavior for users that are accustomed to the current range, especially because we don't have a color-coded legend in the app, and the same range has been used for years. The 2) change will be more subtle, as I'm guessing few users are switching between SNR and C/N0.

See https://github.com/barbeau/gpstest/issues/76#issuecomment-314295438 for details.