fixes #324
Don't let the user pick because we intent him to use email and it wont work if he selects a non-email messenger app. Take the default email app.

Yes, this sucks for people with multiple email clients, but if we don;t do this much of the feedback goes lost because users can't figure out how to send it. Android doesn't have a way that I know of of letting the user choose between only email capable apps.