Make sure user has observed first fix on Status screen before upload
DeviceInfoViewModel is still largely updated in GnssStatusFragment, so we need
to check and make sure that the Status screen has been viewed even if we have a location
to ensure we capture dual-frequency capability, supported GNSS, etc.
Future work should move DeviceInfoViewModel so it's contained in the Activity instead
so no matter what fragment is visible all the DeviceInfoViewModel are still updated.