fix: Allow Status card text to scroll horizontally
This allows the user to scroll the card contents if the text overflows the right side of the screen, as is the case with users that use large fonts or devices with very narrow screens.

Unfortunately this breaks the ripple effect when tapping on the top card to copy the location to the clipboard :(, but the functionality to scroll is more important.

Closes #467