feat: Use viewModel for Sky fragment, including filter
Closes #563

Also bumps a number of libraries.

I hit a strange issue where the filter composable didn't seem to re-compose in Sky the same way it does in Status, presumably because it's mixed into an XML layout, when checking if the filter was active. It would only evaluate the IF statement once and then no longer recompose.

I got around this by putting a State observable first in the IF statement, which presumably avoids some type of optimization that's killing it the other way around:

 if (allStatuses.isNotEmpty() && gnssFilter().isNotEmpty()) {